Rename fetcher to Fetcher
